Bug 736393 - factor out ServerLocalSynchronizer and ServerLocalSynchronizerSession to handle errors in this scenario.
