+++ src/main/java/org/mozilla/gecko/sync/repositories/android/BookmarksDeletionManager.java
@@ -12,7 +12,6 @@ import org.mozilla.gecko.db.BrowserContract;
 import org.mozilla.gecko.sync.Logger;
 import org.mozilla.gecko.sync.repositories.NullCursorException;
 import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionStoreDelegate;
-import org.mozilla.gecko.sync.repositories.domain.BookmarkRecord;
 
 /**
  * Queue up deletions. Process them at the end.
