+++ src/test/java/org/mozilla/android/sync/test/TestServer11RepositorySession.java
@@ -33,6 +33,7 @@ import org.mozilla.gecko.sync.repositories.StoreFailedException;
 import org.mozilla.gecko.sync.repositories.domain.BookmarkRecord;
 import org.mozilla.gecko.sync.repositories.domain.BookmarkRecordFactory;
 import org.mozilla.gecko.sync.repositories.domain.Record;
+import org.mozilla.gecko.sync.synchronizer.ServerLocalSynchronizer;
 import org.mozilla.gecko.sync.synchronizer.Synchronizer;
 import org.simpleframework.http.ContentType;
 import org.simpleframework.http.Request;
@@ -140,13 +141,13 @@ public class TestServer11RepositorySession implements CredentialsSource {
     Crypto5MiddlewareRepository cryptoRepo = new Crypto5MiddlewareRepository(remote, collectionKey);
     cryptoRepo.recordFactory = new BookmarkRecordFactory();
 
-    final Synchronizer synchronizer = new Synchronizer();
+    final Synchronizer synchronizer = new ServerLocalSynchronizer();
     synchronizer.repositoryA = cryptoRepo;
     synchronizer.repositoryB = local;
 
     data.startHTTPServer(server);
     try {
-      Exception e = TestSynchronizer.doSynchronize(synchronizer);
+      Exception e = TestServerLocalSynchronizer.doSynchronize(synchronizer);
       return e;
     } finally {
       data.stopHTTPServer();
